auth users see all workouts

+++ frontend/src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
 
+import { useAuthContext } from '../hooks/useAuthContext'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 
 //date fns
@@ -7,11 +8,16 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({workout}) =>{
     const {dispatch} = useWorkoutsContext()
+    const {user} = useAuthContext()
 
 
     const handleClick = async () =>{
+        if(!user){
+            return
+        }
         const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {'Authorization': `Bearer ${user.token}`}
         })
         const json = await response.json()
